requirements.txt
Django>=3.1.6,<3.2.0
djangorestframework>=3.12.2,<3.13.0
django-rest-swagger>=2.2.0,<2.3.0
psycopg2>=2.7.5,<2.8.0
djangorestframework-simplejwt>=4.6.0,<4.7.0

flake8>=3.6.0,<3.7.0